+++ Anchor.toml
@@ -11,18 +11,13 @@ build = "fkiWK1Wn6ouGcHb3icX4XGKynef5MpsTQ478ZMdgB1g"
 claim_loot = "4CjxHvNUpoCYomULBFTvmkTQPaNd9QDHPhZQ6eB9bZEf"
 claim_time = "4XXA1mX5aN4Fd62FBgNxCU7FzKDYS3KSxFX3RdJYoWPj"
 deity-bot = "62f9zAUjCN5VFqWF43qSUrW6CvivqhsEjDvCHwQ1SjgR"
-location_allocator = "J7q3dEg2KauPKkMamH9Q5FHhCoFYsSq9ramdutMpPTDc"
-loot_distribution = "5F9tMTcNhgjL3tWCaF5HwLkQP9z4XJ4nTXmbYeS8UXRW"
 move_hero = "6o9i5V3EvT9oaokbcZa7G92DWHxcqJnjXmCp94xxhQhv"
 repair = "5xPJt6GDcmGphNAs6qU3hWAvzLwXuSqhTco6RtoAR9aY"
 research = "3ZJ7mgXYhqQf7EsM8q5Ea5YJWA712TFyWGvrj9mRL2gP"
 reset = "3VEXJoAZkYxDXigSWso8FnJY8z6C6inpPxU798vqc9um"
 sacrifice = "6JwZJNAtkciXVGenFSoa99VBNcxyb2W8mvzcMK1vTWKs"
-smart-object-deity = "9RfzWgEBYQAM64a46V3dGRPKYsVY8a7YvZszWPMxvBfk"
 smart-object-deity-interact = "2QPK685TLL7jUG4RYuWXZjv3gw88kUPYw7Aye63cTTjB"
 smart-object-init = "64Uk4oF6mNyviUdK2xHXE3VMCtbCMDgRr1DMJk777DJZ"
-smart-object-location = "5ewDDvpaTkYvoE7ZJJ9cDmZuqvGQt65hsZSJ9w73Fzr1"
-smart-object-token-launcher = "8va4yKEBACkT49C9wo94gS8ZaTdUrq2ipLgZvSNxWbd3"
 smart-object-token-launcher-init = "AdrPpoYr67ZcDZsQxsPgeosE3sQbZxercbUn8i1dcvap"
 smart-object-token-launcher-interact = "DUW1KczxcpeTEY7j9nkvcuAdWGNWoadTeDBKN5Z9xhst"
 token_minter = "4ZxRnucEWC62kVktmx27cz9d1PzWWNgiZLT5VWFLbfB2"
@@ -41,18 +36,13 @@ claim_time = "4XXA1mX5aN4Fd62FBgNxCU7FzKDYS3KSxFX3RdJYoWPj"
 deity-bot = "62f9zAUjCN5VFqWF43qSUrW6CvivqhsEjDvCHwQ1SjgR"
 ecs-bundle = "Cjca6tWWGx77ki6rRinErcoZJEvJHNxfPaut8DoKJHQ5"
 exchange = "Csna3V2jUMdQEQKUCxLsQEnYThAGPSWcPCxW9vea1S8d"
-location_allocator = "J7q3dEg2KauPKkMamH9Q5FHhCoFYsSq9ramdutMpPTDc"
-loot_distribution = "5F9tMTcNhgjL3tWCaF5HwLkQP9z4XJ4nTXmbYeS8UXRW"
 move_hero = "6o9i5V3EvT9oaokbcZa7G92DWHxcqJnjXmCp94xxhQhv"
 repair = "5xPJt6GDcmGphNAs6qU3hWAvzLwXuSqhTco6RtoAR9aY"
 research = "3ZJ7mgXYhqQf7EsM8q5Ea5YJWA712TFyWGvrj9mRL2gP"
 reset = "3VEXJoAZkYxDXigSWso8FnJY8z6C6inpPxU798vqc9um"
 sacrifice = "6JwZJNAtkciXVGenFSoa99VBNcxyb2W8mvzcMK1vTWKs"
-smart-object-deity = "9RfzWgEBYQAM64a46V3dGRPKYsVY8a7YvZszWPMxvBfk"
 smart-object-deity-interact = "2QPK685TLL7jUG4RYuWXZjv3gw88kUPYw7Aye63cTTjB"
 smart-object-init = "64Uk4oF6mNyviUdK2xHXE3VMCtbCMDgRr1DMJk777DJZ"
-smart-object-location = "5ewDDvpaTkYvoE7ZJJ9cDmZuqvGQt65hsZSJ9w73Fzr1"
-smart-object-token-launcher = "8va4yKEBACkT49C9wo94gS8ZaTdUrq2ipLgZvSNxWbd3"
 smart-object-token-launcher-init = "AdrPpoYr67ZcDZsQxsPgeosE3sQbZxercbUn8i1dcvap"
 smart-object-token-launcher-interact = "DUW1KczxcpeTEY7j9nkvcuAdWGNWoadTeDBKN5Z9xhst"
 token_minter = "4ZxRnucEWC62kVktmx27cz9d1PzWWNgiZLT5VWFLbfB2"
@@ -68,7 +58,7 @@ cluster = "localnet"
 wallet = "~/.config/solana/id.json"
 
 [workspace]
-members = ["programs-ecs/components/*", "programs-ecs/systems/*", "programs-ecs/bundles/*", "programs/*"]
+members = ["programs-ecs/systems/*", "programs-ecs/bundles/*", "programs/*"]
 
 [scripts]
 test = "yarn run ts-mocha -p ./tsconfig.json -t 1000000 tests/**/*.test.ts"
