+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-    "programs-ecs/components/*",
     "programs-ecs/systems/*",
     "programs-ecs/bundles/*",
     "programs/*"
