+++ programs-ecs/bundles/ecs-bundle/src/location_allocator/mod.rs
@@ -1,0 +1,1 @@
+pub use crate::ecs_bundle::{LocationAllocator, LocationAllocatorInit};
