+++ programs-ecs/bundles/ecs-bundle/src/smart_object_location/mod.rs
@@ -1,0 +1,1 @@
+pub use crate::ecs_bundle::{SmartObjectLocation, SmartObjectLocationInit};
