--- programs-ecs/components/location_allocator/src/lib.rs
@@ -1,11 +1,0 @@
-use bolt_lang::*;
-
-declare_id!("J7q3dEg2KauPKkMamH9Q5FHhCoFYsSq9ramdutMpPTDc");
-
-#[component(delegate)]
-#[derive(Default)]
-pub struct LocationAllocator {
-    pub current_x: i16,
-    pub current_y: i16,
-    pub direction: u8,
-}
