--- programs-ecs/components/smart-object-deity/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "smart-object-deity"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "smart_object_deity"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build"]
-anchor-debug = []
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
