--- programs-ecs/components/smart-object-location/Xargo.toml
@@ -1,2 +1,0 @@
-[target.bpfel-unknown-unknown.dependencies.std]
-features = []
