--- programs-ecs/components/smart-object-token-launcher/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "smart-object-token-launcher"
-version = "0.1.8"
-description = "Created with Bolt"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib", "lib"]
-name = "smart_object_token_launcher"
-
-[features]
-no-entrypoint = []
-no-idl = []
-no-log-ix-name = []
-cpi = ["no-entrypoint"]
-default = []
-idl-build = ["anchor-lang/idl-build"]
-
-[dependencies]
-bolt-lang.workspace = true
-anchor-lang.workspace = true
-ecs-bundle.workspace = true
