--- programs-ecs/components/smart-object-token-launcher/Xargo.toml
@@ -1,2 +1,0 @@
-[target.bpfel-unknown-unknown.dependencies.std]
-features = []
