+++ programs-ecs/systems/bump_location_allocator/Cargo.toml
@@ -20,4 +20,4 @@ idl-build = ["anchor-lang/idl-build"]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-location_allocator = { version = "0.1.8", path = "../../components/location_allocator", features = ["cpi"] }
+ecs-bundle.workspace = true
