+++ programs-ecs/systems/smart-object-deity-interact/Cargo.toml
@@ -20,7 +20,6 @@ idl-build = ["anchor-lang/idl-build",  "anchor-spl/idl-build"]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-
+ecs-bundle.workspace = true
 deity-bot = { version = "0.1.0", path = "../../../programs/deity-bot", features = ["cpi"] }
-smart-object-deity = { version = "0.1.8", path = "../../components/smart-object-deity", features = ["cpi"] }
 anchor-spl.workspace = true
