+++ programs-ecs/systems/smart-object-init/Cargo.toml
@@ -20,4 +20,4 @@ idl-build = ["anchor-lang/idl-build"]
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-smart-object-location = { version = "0.1.8", path = "../../components/smart-object-location", features = ["cpi"] }
+ecs-bundle.workspace = true
