+++ programs-ecs/systems/smart-object-token-launcher-init/Cargo.toml
@@ -21,5 +21,4 @@ bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
 anchor-spl.workspace = true
-smart-object-token-launcher = { version = "0.1.8", path = "../../components/smart-object-token-launcher", features = ["cpi"] }
 ecs-bundle.workspace = true
