+++ programs-ecs/systems/smart-object-token-launcher-init/src/lib.rs
@@ -15,7 +15,7 @@ pub mod smart_object_token_launcher_init {
         token::{self, spl_token::instruction::AuthorityType, SetAuthority},
     };
     use ecs_bundle::settlement::ResourceBalance;
-    use smart_object_token_launcher::SmartObjectTokenLauncher;
+    use ecs_bundle::smart_object_token_launcher::SmartObjectTokenLauncher;
 
     pub fn execute(
         ctx: Context<Components>,
