+++ programs-ecs/systems/smart-object-token-launcher-interact/Cargo.toml
@@ -21,6 +21,5 @@ anchor-debug = []
 bolt-lang.workspace = true
 anchor-lang.workspace = true
 serde.workspace = true
-smart-object-token-launcher = { version = "0.1.8", path = "../../components/smart-object-token-launcher", features = ["cpi"] }
 ecs-bundle.workspace = true
 anchor-spl.workspace = true
