+++ programs-ecs/systems/smart-object-token-launcher-interact/src/lib.rs
@@ -12,7 +12,7 @@ pub mod smart_object_token_launcher_interact {
     use anchor_spl::token::{mint_to, spl_token, MintTo};
     use bolt_lang::solana_program::program_pack::Pack;
     use ecs_bundle::hero::Hero;
-    use smart_object_token_launcher::SmartObjectTokenLauncher;
+    use ecs_bundle::smart_object_token_launcher::SmartObjectTokenLauncher;
 
     pub fn execute(
         ctx: Context<Components>,
